+++ src/collective/sidebar/upgrades.py
@@ -33,3 +33,8 @@ def to_1300(context):
 def to_1400(context):
     context.runAllImportStepsFromProfile('profile-collective.sidebar:to_1400')
     logger.info('Added collective.sidebar registry records')
+
+
+def to_1500(context):
+    context.runAllImportStepsFromProfile('profile-collective.sidebar:to_1500')
+    logger.info('Added collective.sidebar registry records')